Remove the unused routing_table module
